day 15
This was my worst day so far. Maybe not ever, but so many mistakes just piled up.
First, I did not parse the - of the negative number in the input. And it took a while to catch on.
Then I retrieved the y coordinate for Part 2 and tested the code on the small grid size of the example input.
Guess what var I re-defined and never switched back.
Unlucky.